Move KeyedArrayAccessInterface to Context (#468)
