+++ src/Context/Propagation/KeyedArrayAccessInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\SDK\Trace;
+namespace OpenTelemetry\Context\Propagation;
 
 use ArrayAccess;
